requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
more-itertools==5.0.0
networkx==2.2
numpy==1.16.0
pep8==1.7.1
Pillow==5.4.1
pluggy==0.8.1
py==1.7.0
pydicom==1.2.2
pylint==2.2.2
pyparsing==2.3.1
pytest==4.1.1
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.2
scipy==1.2.1
sfcpy==1.2.1
six==1.12.0
toolz==0.9.0
wrapt==1.11.1
